Ekanayake_IT24100917/aiml/coffee_product_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlalchemy==2.0.36
httpx==0.27.2
pytest==8.3.3